% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.biplot.R
\name{pca.biplot}
\alias{pca.biplot}
\title{pca.biplot}
\usage{
pca.biplot(
  data,
  center = TRUE,
  scale = FALSE,
  pcx = 1,
  pcy = 2,
  groups = 1,
  col,
  pch = 1,
  ellipse = F,
  ellipse.alpha = 0.2,
  ellipse.conf = 0.68,
  ellipse.lwd = 2,
  arrow.options,
  label.options,
  abbrev = F,
  circle = TRUE,
  circle.options,
  legend = TRUE,
  legend.pos = "bottomleft",
  ...
)
}
\arguments{
\item{data}{The data used to build a biplot}

\item{center}{Whether the variables should be shifted to be zero centered (default = TRUE)}

\item{scale}{Whether the variables should be scaled to have unit variance (default = FALSE)}

\item{pcx}{Which principal component to plot on the X-axis (default = 1)}

\item{pcy}{Which principal component to plot on the Y-axis (default = 2)}

\item{groups}{Color observations based on the unique values of this group}

\item{col}{Vector of colors to color observations, needs to be equal to the number of unique values of 'groups' variable, if no 'groups' is defined then all points will have the defined color}

\item{pch}{Determines appearance plotting point, see ?par() for more}

\item{ellipse}{Whether to add confidence ellipses to groups}

\item{ellipse.alpha}{Alpha transparency value for the ellipse (0 = transparent, 1 = opaque, default = 0.2)}

\item{ellipse.conf}{Confidence limit for ellipses}

\item{ellipse.lwd}{Line width of the ellipse}

\item{arrow.options}{List to define options for the variable arrows (for all options see ?arrows ). Needs to be defined as a list, e.g. arrows.options = list(angle = 0.15, col = "blue"), errors could occur if a vector is used.}

\item{label.options}{List to define options for the variable labels (for all options see ?text ). Needs to be defined as a list, e.g. label.options = list(cex = 1.2, col = "red"), errors could occur if a vector is used.}

\item{abbrev}{Whether to abbreviate the variable names}

\item{circle}{Whether to add a unit circle. Only applies if data is scaled.}

\item{circle.options}{List to define options for the unit circle (for all options see ?polygon ). Needs to be defined as a list, e.g. circle.options = list(lty = "dashed", col = "grey")}

\item{legend}{Whether to add a legend}

\item{legend.pos}{Position of the legend}

\item{...}{Other graphical parameters, see ?plot for more details.}
}
\description{
A customizable correlational biplot function. PC's are scaled to unit variance.
}
\examples{
  data(iris)
  pca.biplot(iris[1:4])
  pca.biplot(iris[,-5], col = c("#00AFBB", "#E7B800", "#FC4E07"), groups = iris$Species, 
       legend = F, arrow.options = list(col="blue", angle = 20, length = 0.15))
  pca.biplot(iris[,-5], scale = T, col = c("#00AFBB", "#E7B800", "#FC4E07"), groups = iris$Species, 
       ellipse = T, pch = 16, legend.pos = 'topright')
}
