% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.biplot3d.R
\name{pca.biplot3d}
\alias{pca.biplot3d}
\title{pca.biplot3d}
\usage{
pca.biplot3d(
  data,
  center = TRUE,
  scale = FALSE,
  pcx = 1,
  pcy = 2,
  pcz = 3,
  col = "black",
  groups = NULL,
  bg = "white",
  type = "sphere",
  r = 0.05,
  aspect3d = c(1, 1, 1),
  abbrev = FALSE,
  arrow.col = "black",
  label.col = "darkblue",
  label.cex = 1,
  simple.axis = FALSE,
  box = FALSE,
  grid = FALSE,
  ellipse = FALSE,
  ellipse.conf = 0.68,
  ellipse.alpha = 0.2,
  ellipse.shade = TRUE,
  ellipse.wire = FALSE,
  ...
)
}
\arguments{
\item{data}{The data used to build a biplot}

\item{center}{Whether the variables should be shifted to be zero centered (default = TRUE)}

\item{scale}{Whether the variables should be scaled to have unit variance (default = FALSE)}

\item{pcx}{Which principal component to plot on the X-axis (default = 1)}

\item{pcy}{Which principal component to plot on the Y-axis (default = 2)}

\item{pcz}{Which principal component to plot on the Z-axis (default = 3)}

\item{col}{Change color of observations, only if no group is defined}

\item{groups}{Color observations based on the unique values of this group}

\item{bg}{Background color}

\item{type}{Type of plot. Options: "sphere", "point", "text" (default = "sphere")}

\item{r}{Vector or single value defining the sphere radius/radii}

\item{aspect3d}{Vector used to set apparent ratios of the x, y, and z axes of the current bounding box}

\item{abbrev}{Whether to abbreviate the variable names}

\item{arrow.col}{Color of the variable arrow}

\item{label.col}{Color of the variable label}

\item{label.cex}{Size of the variable label}

\item{simple.axis}{Use axes3d instead of decorate3d if TRUE}

\item{box}{Whether to draw a box (only if using decorate3d)}

\item{grid}{Adds a reference grid}

\item{ellipse}{Whether to add confidence ellipses to groups (to display this on the plot, ellipse.shade and/or ellipse.wire must be TRUE)}

\item{ellipse.conf}{Confidence limit for ellipses}

\item{ellipse.alpha}{Alpha transparency value for the ellipse (0 = transparent, 1 = opaque, default = 0.2)}

\item{ellipse.shade}{Draws ellipse as a surface}

\item{ellipse.wire}{Draws ellipse as line segments}

\item{...}{Material properties (see rgl.material).}
}
\description{
An interactive 3D biplot. PC's are scaled to unit variance.
}
\examples{
     data(iris)
     pca.biplot3d(iris[,-5], scale = T)
     pca.biplot3d(iris[,-5], groups = iris$Species, scale = T, box = T, grid = T, ellipse = T)

}
