% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.screeplot.R
\name{pca.screeplot}
\alias{pca.screeplot}
\title{pca.screeplot}
\usage{
pca.screeplot(data, center = T, scale = F, type = 1)
}
\arguments{
\item{data}{The data used to build the screeplot}

\item{center}{Center data before building screeplot}

\item{scale}{Scale data before building screeplot}

\item{type}{Type of screeplot (1 = dots and lines, 2 = barplot)}
}
\description{
Plots the percentage of explained variance against the number of the principal component.
}
\examples{
     data(iris)
     pca.screeplot(iris[,-5], scale = T)

}
