% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.biplot.R
\name{pca.biplot}
\alias{pca.biplot}
\title{pca.biplot}
\usage{
pca.biplot(
  data,
  center = TRUE,
  scale = FALSE,
  pcx = 1,
  pcy = 2,
  groups = 1,
  col = NULL,
  pch = 1,
  ellipse = FALSE,
  ellipse.alpha = 0.2,
  ellipse.conf = 0.68,
  ellipse.lwd = 2,
  arrow.options,
  label.options,
  abbrev = FALSE,
  circle = TRUE,
  circle.options,
  legend = TRUE,
  legend.pos = "bottomleft",
  ...
)
}
\arguments{
\item{data}{The data used to build the biplot}

\item{center}{Column centering the data (default = `TRUE`). If center is `TRUE` then centering is done by subtracting the column means (omitting `NAs`) of x from their corresponding columns, and if center is `FALSE`, no centering is done.}

\item{scale}{Scaling the data (default = `FALSE`). If scale is `TRUE` then scaling is done by dividing the (centered) columns of x by their standard deviations if center is `TRUE`, and the root mean square otherwise. If scale is `FALSE`, no scaling is done.}

\item{pcx}{Which principal component to plot on the X-axis (default = 1)}

\item{pcy}{Which principal component to plot on the Y-axis (default = 2)}

\item{groups}{A vector from which the unique values are used to color the observations}

\item{col}{Vector of colors, to color observations. Needs to be equal to the number of unique values of `groups` variable. If no `groups` is defined then all points will have the defined color.}

\item{pch}{Either an integer specifying a symbol or a single character to be used as the default in plotting points. See [graphics::points()] for possible values and their interpretation.}

\item{ellipse}{Whether to add confidence ellipses to the different groups}

\item{ellipse.alpha}{Alpha transparency value for the fill color (0 = transparent, 1 = opaque)}

\item{ellipse.conf}{Confidence limit for ellipses, e.g. 0.68. If given, the corresponding sd or se is multiplied with the corresponding value found from the Chi-squared distribution with 2df.}

\item{ellipse.lwd}{Line width of ellipse border}

\item{arrow.options}{List to define options for the variable arrows (for all options see [graphics::arrows()] ). Needs to be defined as a list, e.g. `arrows.options = list(angle = 0.15, col = "blue")`, errors could occur if a vector is used.}

\item{label.options}{List to define options for the variable labels (for all options see [graphics::text()] ). Needs to be defined as a list, e.g. `label.options = list(cex = 1.2, col = "red")`, errors could occur if a vector is used.}

\item{abbrev}{Whether to abbreviate the variable names}

\item{circle}{Whether to add a unit circle. Only applies if data is scaled!}

\item{circle.options}{List to define options for the unit circle (for all options see [graphics::polygon()] ). Needs to be defined as a list, e.g. `circle.options = list(lty = "dashed", col = "grey")`, errors could occur if a vector is used.}

\item{legend}{Whether to add a legend}

\item{legend.pos}{Position of the legend}

\item{...}{Other graphical parameters, see [graphics::plot()] for more details.}
}
\description{
A customizable correlational biplot function. PC's are scaled to unit variance.
}
\examples{
  data(iris)
  pca.biplot(iris[1:4])
  pca.biplot(iris[,-5], col = c("#00AFBB", "#E7B800", "#FC4E07"), groups = iris$Species, 
       legend = F, arrow.options = list(col="blue", angle = 20, length = 0.15))
  pca.biplot(iris[,-5], scale = T, col = c("#00AFBB", "#E7B800", "#FC4E07"), groups = iris$Species, 
       ellipse = T, pch = 16, legend.pos = 'topright')
}
