% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.screeplot.R
\name{pca.screeplot}
\alias{pca.screeplot}
\title{pca.screeplot}
\usage{
pca.screeplot(data, center = T, scale = F, type = 1)
}
\arguments{
\item{data}{The data used to build the screeplot}

\item{center}{Column centering the data (default = `TRUE`). If center is `TRUE` then centering is done by subtracting the column means (omitting `NAs`) of x from their corresponding columns, and if center is `FALSE`, no centering is done.}

\item{scale}{Scaling the data (default = `FALSE`). If scale is `TRUE` then scaling is done by dividing the (centered) columns of x by their standard deviations if center is `TRUE`, and the root mean square otherwise. If scale is `FALSE`, no scaling is done.}

\item{type}{Type of screeplot (1 = dots and lines, 2 = barplot)}
}
\description{
Plots the percentage of explained variance against the number of the principal component.
}
\examples{
     data(iris)
     pca.screeplot(iris[,-5], scale = T)

}
